[ refactor ] 컨벤션에 맞게 컴포넌트 변경
